+++ app/PersonioIntegration/Widgets/Single.php
@@ -1,0 +1,102 @@
+<?php
+/**
+ * File to handle the single widget.
+ *
+ * @package personio-integration-light
+ */
+
+namespace PersonioIntegrationLight\PersonioIntegration\Widgets;
+
+// prevent direct access.
+defined( 'ABSPATH' ) || exit;
+
+use PersonioIntegrationLight\PersonioIntegration\Widget_Base;
+
+/**
+ * Object to handle the description widget.
+ */
+class Single extends Widget_Base {
+	/**
+	 * Internal name for this object.
+	 *
+	 * @var string
+	 */
+	protected string $name = 'widget_single';
+
+	/**
+	 * Name if the setting field which defines its state.
+	 *
+	 * @var string
+	 */
+	protected string $setting_field = 'personioIntegrationWidgetSingleStatus';
+
+	/**
+	 * Name if the setting tab where the setting field is visible.
+	 *
+	 * @var string
+	 */
+	protected string $setting_tab = '';
+
+	/**
+	 * Path to Block object.
+	 *
+	 * @var string
+	 */
+	protected string $gutenberg = '\PersonioIntegrationLight\PageBuilder\Gutenberg\Blocks\Single';
+
+	/**
+	 * Instance of this object.
+	 *
+	 * @var ?Single
+	 */
+	private static ?Single $instance = null;
+
+	/**
+	 * Return the instance of this Singleton object.
+	 */
+	public static function get_instance(): Single {
+		if ( is_null( self::$instance ) ) {
+			self::$instance = new self();
+		}
+
+		return self::$instance;
+	}
+
+	/**
+	 * Return label of this extension.
+	 *
+	 * @return string
+	 */
+	public function get_label(): string {
+		return __( 'Single', 'personio-integration-light' );
+	}
+
+	/**
+	 * Return whether this extension is enabled (true) or not (false).
+	 *
+	 * @return bool
+	 */
+	public function is_enabled(): bool {
+		return 1 === absint( get_option( $this->get_settings_field_name() ) );
+	}
+
+	/**
+	 * Return the description for this extension.
+	 *
+	 * @return string
+	 */
+	public function get_description(): string {
+		return __('Provides a widget to show a single Personio position.', 'personio-integration-light');
+	}
+
+	/**
+	 * Return the rendered widget.
+	 *
+	 * @param array<string,mixed> $attributes Attributes to configure the rendering.
+	 *
+	 * @return string
+	 */
+	public function render( array $attributes ): string {
+		return Archive::get_instance()->render( $attributes );
+	}
+}