+++ changelog.md
@@ -15,6 +15,8 @@
 - Added new statistic about the plugin data, which could also be sent via email on regular base
 - Added option to change the from-email in each email
 - Added email-template for all emails
+- New centralized widget handling for every supported PageBuilder
+- Added new extension category "Widgets"
 - Added hint for additional offices which are usable in Pro-plugin
 
 ### Changed